requirements.txt
absl-py==2.1.0
accelerate==0.24.1
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
antlr4-python3-runtime==4.8
anyio==4.3.0
argcomplete==3.3.0
argparse==1.4.0
asteroid==0.7.0
asteroid-filterbanks==0.4.0
async-timeout==4.0.3
attrs==23.2.0
audiomentations==0.35.0
audioread==3.0.1
audiosr==0.0.7
auraloss==0.4.0
babel==2.15.0
beartype==0.19.0
bibtexparser==2.0.0b7
bitarray==2.9.2
black==24.1.1
braceexpand==0.1.7
byol-pytorch==0.8.2
cached-property==1.5.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
clldutils==3.22.2
cmake==3.29.3
colorama==0.4.6
colorlog==6.8.2
contourpy==1.2.1
csvw==3.3.0
cycler==0.12.1
cython==3.0.10
cytoolz==0.12.3
datasets==2.20.0
debugpy==1.8.1
decorator==5.1.1
diffsptk==1.0.1
diffusers==0.28.0.dev0
dill==0.3.8
distance==0.1.3
distlib==0.3.8
dlinfo==1.2.1
dnspython==2.6.1
docker-pycreds==0.4.0
easydict==1.13
einops==0.8.0
einx==0.2.2
email-validator==2.1.1
encodec==0.1.1
exceptiongroup==1.2.1
fairseq==0.12.2
fastapi==0.111.0
fastapi-cli==0.0.4
fastdtw==0.3.4
ffmpeg-python==0.2.0
ffmpy==0.3.2
filelock==3.14.0
fonttools==4.51.0
frechet-audio-distance==0.3.1
frozendict==2.4.4
frozenlist==1.4.1
fsspec==2024.3.1
ftfy==6.2.0
future==1.0.0
g2p-en==2.1.0
gitdb==4.0.11
gitpython==3.1.43
gradio==4.44.1
gradio-client==1.3.0
grpcio==1.63.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.5
humanfriendly==10.0
hydra-core==1.0.7
idna==3.7
importlib-metadata==7.1.0
importlib-resources==6.4.0
inflect==7.2.1
iniconfig==2.0.0
intervaltree==3.1.0
isodate==0.6.1
jinja2==3.1.4
jiwer==3.0.5
joblib==1.4.2
json5==0.9.25
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
julius==0.2.7
kiwisolver==1.4.5
laion-clap==1.1.2
language-tags==1.2.0
lhotse==1.24.0.dev0+git.ddde5bd.clean
librosa==0.9.2
lightning-utilities==0.11.2
lilcom==1.7
lit==18.1.4
llvmlite==0.42.0
local-attention==1.9.3
loguru==0.7.2
lxml==5.2.2
markdown==3.6
markdown-it-py==3.0.0
markupsafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
mir-eval==0.7
modelscope==1.21.0
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
networkx==3.2.1
ninja==1.11.1.1
nltk==3.8.1
nnaudio==0.3.3
nox==2024.4.15
numba==0.59.1
numpy==1.23.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-ml-py==12.535.161
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
nvitop==1.3.2
omegaconf==2.0.6
openai-whisper==20231117
orjson==3.10.3
packaging==24.0
pandas==2.2.2
pathspec==0.12.1
pb-bss-eval==0.0.2
pedalboard==0.9.5
peft==0.12.0
pesq==0.0.4
phonemizer==3.2.1
pillow==10.3.0
pip==24.0
platformdirs==4.2.1
pluggy==1.5.0
pooch==1.8.1
portalocker==2.8.2
praat-parselmouth==0.4.3
progressbar==2.5
protobuf==4.25.3
psutil==5.9.8
ptwt==0.1.8
pyarrow==16.1.0
pyarrow-hotfix==0.6
pycparser==2.22
pydantic==2.7.1
pydantic-core==2.18.2
pydub==0.25.1
pygments==2.18.0
pylatexenc==2.10
pymcd==0.2.1
pyparsing==3.1.2
pypesq==1.2.4
pypinyin==0.48.0
pysptk==0.2.2
pystoi==0.4.1
pytest==8.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytorch-lightning==2.2.4
pytorch-msssim==1.0.0
pytorch-ranger==0.1.1
pytz==2024.1
pywavelets==1.6.0
pyworld==0.3.4
pyyaml==6.0.1
rapidfuzz==3.10.1
rdflib==7.0.0
referencing==0.35.1
regex==2024.5.10
requests==2.32.3
resampy==0.4.3
resemblyzer==0.1.4
rfc3986==1.5.0
rich==13.7.1
rpds-py==0.18.1
ruamel-yaml==0.18.6
ruamel-yaml-clib==0.2.8
ruff==0.4.5
sacrebleu==2.4.2
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.10.1
segments==2.2.1
semantic-version==2.10.0
sentry-sdk==2.1.1
setproctitle==1.3.3
setuptools==69.5.1
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.3.7
starlette==0.37.2
svgwrite==1.4.3
sympy==1.12
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardx==2.6.2.2
termcolor==2.4.0
tgt==1.5
threadpoolctl==3.5.0
tiktoken==0.6.0
timm==1.0.3
tokenizers==0.13.3
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.0.1+cu118
torch-optimizer==0.1.0
torch-stoi==0.2.1
torchaudio==2.0.2+cu118
torchcrepe==0.0.18
torchlibrosa==0.1.0
torchlpc==0.4
torchmetrics==0.11.4
torchvision==0.15.2+cu118
tqdm==4.66.4
transformers==4.30.2
tree==0.2.4
triton==2.0.0
typeguard==4.2.1
typer==0.12.3
typing==3.7.4.3
typing-extensions==4.11.0
tzdata==2024.1
ujson==5.10.0
unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
vector-quantize-pytorch==1.12.5
virtualenv==20.26.1
wandb==0.17.0
watchfiles==0.21.0
wcwidth==0.2.13
webdataset==0.2.86
webrtcvad==2.0.10
websockets==11.0.3
werkzeug==3.0.3
wget==3.2
wheel==0.43.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.18.1
